requirements.txt
# Web Framework
fastapi
uvicorn
pydantic

# Database
sqlalchemy
psycopg2-binary

# IoT & Async
paho-mqtt
asyncio

# Data Science (Let pip choose the best version)
numpy
pandas
scikit-learn
joblib

# Testing
pytest
httpx

minio
redis
tensorflow